x64 trampoline: ensure stack is 16 bytes aligned
RSP is not 16 bytes aligned in body of Trampoline_ASM_x64,
after the following instruction:

    sub rsp, 32; shadow space for method calls

This results in crashes if GFlags and PageHeap is used.
The ASM code with misaligned RSP calls LhSetInclusiveACL which indirectly
calls RtlCaptureContext, which crahes in ntdll!RtlCaptureContext+0x86
fxsave which in turn is called with misaligned RSP.

It was suggested to adjust ASM code to ensure RSP is aligned in
https://github.com/EasyHook/EasyHook/issues/214#issuecomment-348846356

I used technique described in https://stackoverflow.com/a/9600102

    pushq %rsp
    pushq (%rsp)
    andq $-0x10, %rsp
        call function
    movq 8(%rsp), %rsp

> The two pushes leave the stack with the same alignment it had originally, and
> a copy of the original %rsp at (%rsp) and 8(%rsp). The andq then aligns the
> stack - if it was already 16 byte aligned nothing changes, if it was 8 byte
> aligned then it subtracts 8 from %rsp, meaning that the original %rsp is now
> at 8(%rsp) and 16(%rsp). So we can unconditionally restore it from 8(%rsp).

In Trampoline_ASM_x64 I adjusted references to RSP in two places which are
under the alignment trick. Otherwise they refer to modified RSP.

Fix https://github.com/EasyHook/EasyHook/issues/214